fixed bugs and can run vgg+voc+NOeval successfully

+++ layers/modules/multibox_loss.py
@@ -115,7 +115,7 @@ class MultiBoxLoss(nn.Module):
         # Sum of losses: L(x,c,l,g) = (Lconf(x, c) + αLloc(x,l,g)) / N
 
         N = num_pos.data.sum()
-        N = N.float()# for pytorch 0.4
+        #N = N.float()# for pytorch 0.4
         loss_l /= N
         loss_c /= N
         return loss_l, loss_c
